[feat]Layout으로 변경
